Refactored CodeGenDelegationManager package
