+++ src/main/java/com/sachett/slang/slangc/codegen/utils/delegation/CodeGenMethodMap.java
@@ -1,4 +1,4 @@
-package com.sachett.slang.slangc.codegen.utils;
+package com.sachett.slang.slangc.codegen.utils.delegation;
 
 import com.sachett.slang.parser.SlangParser;
 import org.antlr.v4.runtime.tree.ParseTree;
